made animation on new post properly working and made it not run twice on my own submit
